+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `Unit` and `UnitBuilder` traits.
 - Remove `Layout` trait.
 - Remove `func_layout()` and `func_layout_mut()` functions.
+- Remove `cfg()`, `cfg_mut()`, `try_cfg()`, and `try_cfg_mut()` functions.
 
 ## 0.12.0 - 2020-04-09
 ### Added